+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- CLI command block in notebook shows full command path
+
 ## [0.11.5] - 2026-06-10
 
 ### Fixed