+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.12.0] - 2026-06-11
+
 ### Added
 
 - `moops.embed` accepts a `keep=` sequence of definition names to retain in